Workaround for Issue #95: Calls to UpdateDefaultConfigFile() in UE4.23 does not to a diff as expected.
Call incorrectly regenerates axis properties in random order each time. although values are accurate, it does trigger diffs for Version Control due to random line placement of the entries.

Behavior is not  present in UE4.15-4.22 and also not in upcoming UE4.24, seems solely an issue in UE4.23

Updated plugin to use SaveConfig() instead which does not exhibit this randomness in UE4.23.